Move #withRateLimit extensions to RateLimited
