+++ src/main/kotlin/com/freya02/botcommands/internal/commands/application/ApplicationCommandListener.kt
@@ -12,6 +12,7 @@ import com.freya02.botcommands.internal.ExceptionHandler
 import com.freya02.botcommands.internal.Usability
 import com.freya02.botcommands.internal.Usability.UnusableReason
 import com.freya02.botcommands.internal.commands.application.slash.SlashCommandInfo
+import com.freya02.botcommands.internal.commands.withRateLimit
 import com.freya02.botcommands.internal.utils.throwInternal
 import dev.minn.jda.ktx.messages.reply_
 import kotlinx.coroutines.launch
@@ -40,7 +41,7 @@ internal class ApplicationCommandListener(private val context: BContextImpl) {
                 }
 
                 val isNotOwner = !context.isOwner(event.user.idLong)
-                withRateLimit(event, slashCommand, isNotOwner) {
+                slashCommand.withRateLimit(context, event, isNotOwner) {
                     if (!canRun(event, slashCommand, isNotOwner)) {
                         false
                     } else {
@@ -65,7 +66,7 @@ internal class ApplicationCommandListener(private val context: BContextImpl) {
                 }
 
                 val isNotOwner = !context.isOwner(event.user.idLong)
-                withRateLimit(event, userCommand, isNotOwner) {
+                userCommand.withRateLimit(context, event, isNotOwner) {
                     if (!canRun(event, userCommand, isNotOwner)) {
                         false
                     } else {
@@ -90,7 +91,7 @@ internal class ApplicationCommandListener(private val context: BContextImpl) {
                 }
 
                 val isNotOwner = !context.isOwner(event.user.idLong)
-                withRateLimit(event, messageCommand, isNotOwner) {
+                messageCommand.withRateLimit(context, event, isNotOwner) {
                     if (!canRun(event, messageCommand, isNotOwner)) {
                         false
                     } else {
@@ -100,26 +101,6 @@ internal class ApplicationCommandListener(private val context: BContextImpl) {
             } catch (e: Throwable) {
                 handleException(e, event)
             }
-        }
-    }
-
-    private suspend fun withRateLimit(event: GenericCommandInteractionEvent, command: ApplicationCommandInfo, isNotOwner: Boolean, block: suspend () -> Boolean) {
-        val rateLimitInfo = command.rateLimitInfo
-        if (isNotOwner && rateLimitInfo != null) {
-            val bucket = rateLimitInfo.helper.getBucket(context, event, command)
-            val probe = bucket.tryConsumeAndReturnRemaining(1)
-            if (probe.isConsumed) {
-                try {
-                    block()
-                } catch (e: Throwable) {
-                    bucket.addTokens(1)
-                    throw e
-                }
-            } else {
-                rateLimitInfo.helper.onRateLimit(context, event, command, probe)
-            }
-        } else {
-            block()
         }
     }
